+++ src/main/java/net/tracystacktrace/stackem/processor/iconswap/IconSwapReader.java
@@ -3,6 +3,7 @@ package net.tracystacktrace.stackem.processor.iconswap;
 import com.google.gson.*;
 import net.minecraft.common.item.Items;
 import net.tracystacktrace.stackem.StackEm;
+import net.tracystacktrace.stackem.processor.iconswap.swap.ISwapper;
 import net.tracystacktrace.stackem.processor.iconswap.swap.TextureByMetadata;
 import net.tracystacktrace.stackem.processor.iconswap.swap.TextureByName;
 import net.tracystacktrace.stackem.tools.JsonReadHelper;
@@ -119,14 +120,14 @@ public final class IconSwapReader {
         String texture = null;
         if (object.has("texture")) {
             final JsonElement textureElement = object.get("texture");
-            if(textureElement.isJsonPrimitive() && textureElement.getAsJsonPrimitive().isString()) {
+            if (textureElement.isJsonPrimitive() && textureElement.getAsJsonPrimitive().isString()) {
                 texture = textureElement.getAsString();
             } else {
                 throw new IconProcessorException(IconProcessorException.INVALID_TEXTURE, textureElement.toString());
             }
         }
 
-        if(texture == null) {
+        if (texture == null) {
             throw new IconProcessorException(IconProcessorException.NOT_FOUND_TEXTURE);
         }
 
@@ -134,15 +135,34 @@ public final class IconSwapReader {
     }
 
     public static @Nullable Integer obtainPriority(@NotNull JsonObject object) throws IconProcessorException {
-        if(object.has("priority")) {
+        if (object.has("priority")) {
             final JsonElement element = object.get("priority");
-            if(element.isJsonPrimitive() && element.getAsJsonPrimitive().isNumber()) {
+            if (element.isJsonPrimitive() && element.getAsJsonPrimitive().isNumber()) {
                 return element.getAsInt();
             } else {
                 throw new IconProcessorException(IconProcessorException.INVALID_PRIORITY, element.toString());
             }
         }
         return null;
+    }
+
+    public static void obtainArmorIfPossible(@NotNull ISwapper swapper, @NotNull JsonObject object) throws IconProcessorException {
+        if (object.has("armorTexture")) {
+            final JsonElement element = object.get("armorTexture");
+            if (element.isJsonPrimitive() && element.getAsJsonPrimitive().isString()) {
+                swapper.setArmorTexture(element.getAsString());
+            } else {
+                throw new IconProcessorException(IconProcessorException.INVALID_ARMOR_TEXTURE, element.toString());
+            }
+        }
+        if(object.has("armorEnableColor")) {
+            final JsonElement element = object.get("armorEnableCoor");
+            if(element.isJsonPrimitive() && element.getAsJsonPrimitive().isBoolean()) {
+                swapper.setEnableColor(element.getAsBoolean());
+            } else {
+                throw new IconProcessorException(IconProcessorException.INVALID_ENABLE_COLOR, element.toString());
+            }
+        }
     }
 
     private static boolean isValidItemID(int i) {